Level code cleanup
- Level generation code moved `LevelGenerator.cs`. This way level generation and level logic code are separate. Easier to understand and to find code.
- Different room types/obstacles are implemented by creating a new class which inherits `Room`. See Level.cs line 211.
- Added OnEnter() and OnExit() to Room as an easier way to start/stop room specific sound effects like wind.
- The player instance is now owned by GameLogic instead of Level. The player isn't recreated each time they beat a level anymore. This makes more sense and fixes inventory getting wiped every time a level is beat.
- Score is now preserved when completing a level.
- Keys are correctly removed from the players inventory when completing the end room.
